Suppression de dépendances à un back-end (#51)
* WIP

* feat: Remove unnecessary linear progress bar in App component and update hpHcJson retrieval logic

* feat: Refactor ComparisonTable and Simulations components

Refactor the ComparisonTable component to remove unnecessary comments and improve code readability. Also, update the Simulations component to include media query and theme usage for better responsiveness.

* feat: Refactor build workflows to remove unnecessary directory change

Refactor the build workflows to remove the unnecessary directory change to the 'front' directory. This simplifies the build process and improves code readability.

* chore: Refactor build workflows and update workflow names

Refactor the build workflows to remove unnecessary directory changes and update the workflow names for better clarity and consistency. This simplifies the build process, improves code readability, and aligns with established conventions.

* chore: Refactor build workflows and update workflow names

* feat: Set timezone to "Europe/Paris" for test workflow

* feat: Tests

Refactor the `OptionKey` enum in `types.ts` to update the value of `TEMPO` to `BLEU_TEMPO`. This change aligns with the updated naming convention and improves code clarity.

* sort

* Sonar quality

* feat: Update react-router to version 7.3.0

Update the react-router package to version 7.3.0 to leverage the latest features and improvements.

* coverage?

* chore: Update Sonar coverage exclusions

* feat: Update fetch-and-update-tempo-data script

Add a new npm script `fetch-and-update-tempo-data` to the `package.json` file. This script runs the `updateTempo.ts` script located in the `src/scripts` directory. It fetches the latest tempo data and updates the `tempo.json` file in the `src/assets` directory. This improvement enhances the data fetching process and ensures up-to-date tempo information.

* fs

* sonar exclusions

* exclusion

* remove from front build fs services

* FS DEPCHECK

* react-router-dom

* chore: Refactor fetchAndUpdateTempo.ts and utils.ts

Refactor the `fetchAndUpdateTempo.ts` and `utils.ts` files to improve code organization and remove duplicate code. Move the `fetchTempoData` function from `utils.ts` to `fetchAndUpdateTempo.ts` and remove the duplicate import of `axios` in `utils.ts`. This refactoring enhances code maintainability and reduces redundancy.

* Self review

* review + sonar